modified data type in method (int to double), added StatsServiceTest

+++ src/main/java/ru/netology/stats/StatsService.java
@@ -1,20 +1,20 @@
 package ru.netology.stats;
 
 public class StatsService {
-    public int calculateSum(int[] sales) {
-        int sum = 0;
-        for (int sale : sales) {
+    public double calculateSum(double[] sales) {
+        double sum = 0;
+        for (double sale : sales) {
             sum += sale;
         }
         return sum;
     }
 
-    public int calculateAverage(int[] sales) {
+    public double calculateAverage(double[] sales) {
         return calculateSum(sales) / sales.length;
     }
 
-    public int findMaxMonth(int[] sales) {
-        int maxSale = sales[0];
+    public int findMaxMonth(double[] sales) {
+        double maxSale = sales[0];
         int maxMonth = 0;
         for (int i = 0; i < sales.length; i++) {
             if (sales[i] >= maxSale) {
@@ -25,8 +25,8 @@ public class StatsService {
         return maxMonth + 1; // сдвигаем ответ на +1, т.к. месяц нумеруется с 1,  а массив с 0
     }
 
-    public int findMinMonth(int[] sales) {
-        int minSale = sales[0];
+    public int findMinMonth(double[] sales) {
+        double minSale = sales[0];
         int minMonth = 0;
         for (int i = 0; i < sales.length; i++) {
             if (sales[i] <= minSale) {
@@ -37,10 +37,10 @@ public class StatsService {
         return minMonth + 1; // сдвигаем ответ на +1, т.к. месяц нумеруется с 1,  а массив с 0
     }
 
-    public int calculateMonthsBelowAverage(int[] sales) {
-        int average = calculateAverage(sales);
+    public int calculateMonthsBelowAverage(double[] sales) {
+        double average = calculateAverage(sales);
         int monthsBelowAverage = 0;
-        for (int sale : sales) {
+        for (double sale : sales) {
             if (sale < average) {
                 monthsBelowAverage++;
             }
@@ -48,10 +48,10 @@ public class StatsService {
         return monthsBelowAverage;
     }
 
-    public int calculateMonthsAboveAverage(int[] sales) {
-        long average = calculateAverage(sales);
+    public int calculateMonthsAboveAverage(double[] sales) {
+        double average = calculateAverage(sales);
         int monthsAboveAverage = 0;
-        for (int sale : sales) {
+        for (double sale : sales) {
             if (sale > average) {
                 monthsAboveAverage++;
             }
